// src/services/models/leaderboardControllerGetLeaderboardMetric.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * TickX Trading Core API
 * OpenAPI spec version: 0.0.12
 */

export type LeaderboardControllerGetLeaderboardMetric = typeof LeaderboardControllerGetLeaderboardMetric[keyof typeof LeaderboardControllerGetLeaderboardMetric];


export const LeaderboardControllerGetLeaderboardMetric = {
  pnl: 'pnl',
  volume: 'volume',
} as const;